refactor: 로또 번호가 당첨 번호와 같은지 묻는 주체 변경
인자로 넘겨져오는 LottoTicket에 메시지를 보내서 같은 번호인지를 묻도록 변경
