+++ src/main/java/lotto/domain/LottoTicket.java
@@ -15,4 +15,8 @@ public class LottoTicket {
     public List<Integer> getLottoNumbers() {
         return Collections.unmodifiableList(lottoNumbers);
     }
+
+    public boolean isSame(WinningNumber winningNumber) {
+        return lottoNumbers.contains(winningNumber.getWinningNumber());
+    }
 }
