+++ src/test/java/lotto/domain/WinningNumberTest.java
@@ -20,18 +20,4 @@ class WinningNumberTest {
         // then
         assertThat(result).isTrue();
     }
-
-    @DisplayName("LottoTicket을 기반으로 당첨 번호가 존재하는지 유무를 반환한다.")
-    @Test
-    void 당첨_번호_인지_확인() {
-        // given
-        LottoTicket lottoTicket = new LottoTicket(() -> List.of(1, 2, 3, 4, 5, 6));
-        WinningNumber winningNumber = new WinningNumber(1, BallType.BONUS);
-
-        // when
-        boolean result = winningNumber.isSame(lottoTicket);
-
-        // then
-        assertThat(result).isTrue();
-    }
 }
